Add aliases to allow external programs to run if emulator is MinTTY
Git for Windows recently switched to using MinTTY as the default emulator.
There is a long running issue with running external programs in MinTTY[1]
and one solution is to run programs through winpty. This commit creates
aliases that run the program on its own when possible and fall back to
using winpty when necessary.

[1] https://github.com/mintty/mintty/issues/56